+++ Content.Client/Construction/EntitySystems/AnchorableSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared.Construction.EntitySystems;
+
+namespace Content.Client.Construction.EntitySystems;
+
+public sealed class AnchorableSystem : SharedAnchorableSystem { }
