+++ Content.Client/Entry/IgnoredComponents.cs
@@ -5,7 +5,6 @@ namespace Content.Client.Entry
     {
         public static string[] List => new[]
         {
-            "Anchorable",
             "AmmoBox",
             "Pickaxe",
             "IngestionBlocker",
@@ -89,7 +88,6 @@ namespace Content.Client.Entry
             "SolarPanel",
             "BodyScanner",
             "Stunbaton",
-            "Tool",
             "TilePrying",
             "RandomSpriteColor",
             "ConditionalSpawner",
