+++ Content.Server/Nuke/NukeSystem.cs
@@ -1,10 +1,10 @@
 using Content.Server.Chat.Managers;
-using Content.Server.Construction.Components;
 using Content.Server.Coordinates.Helpers;
 using Content.Server.Popups;
 using Content.Server.UserInterface;
 using Content.Shared.Audio;
 using Content.Shared.Body.Components;
+using Content.Shared.Construction.Components;
 using Content.Shared.Containers.ItemSlots;
 using Content.Shared.Nuke;
 using Content.Shared.Sound;
