+++ Content.Server/Tools/ToolSystem.TilePrying.cs
@@ -4,6 +4,7 @@ using Content.Server.Tools.Components;
 using Content.Shared.Interaction;
 using Content.Shared.Interaction.Helpers;
 using Content.Shared.Maps;
+using Content.Shared.Tools.Components;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Map;
