+++ Content.Server/Weapon/Melee/EnergySword/EnergySwordSystem.cs
@@ -1,8 +1,8 @@
-using Content.Server.Tools.Components;
 using Content.Shared.ActionBlocker;
 using Content.Shared.Interaction;
 using Content.Shared.Interaction.Events;
 using Content.Shared.Item;
+using Content.Shared.Tools.Components;
 using Content.Shared.Weapons.Melee;
 using Robust.Shared.Audio;
 using Robust.Shared.GameObjects;
